A rockin' tan.
- Changed some Tanning recipe mechanics. They now require 4 additional
arguments: consumeItem (boolean that determines whether the "tool" item
will be consumed upon use), sound (the sound that plays upon using the
tool item. Not yet implemented), particle (EnumParticleType that will
display while the item is tanning), and sun (boolean that determines
whether the item must be out in the sun in order to tan).
- Reworked some of the internal logic on Tanning Rack tile entities.
- Same with the block itself.
- There's a lot of new bugs with tanning, I'll be working those out over
the next few days.