TAAAAAANING RAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA
- Updated Tanning Rack's bounding and collision boxes to better match
its model.
- Fixed a bug causing items to face the wrong direction when placed on
Tanning Racks.
- Fixed multiple Dagger types not being usable on Tanning recipes with a
Dagger as their tool.
- Tanning recipes requiring the use of a Dagger now accept any Dagger of
the same harvest level as the one specified in the recipe (ie, if the
recipe calls for Iron, then Steel, Bronze, Meurodite, or Torite will
also suffice).
- Fixed an oversight where the Tanning Rack couldn't handle multiple
recipes with the same item but different tools.
- Tanning Rack now emits smoke particles while in the process of tanning
an item. May change this in the future, but for now I can't seem to get
variable recipe-specified particles to work, so whatever.
- Fixed items being slightly misaligned on Tanning Racks facing west.
- Added Sawdust. Currently obtained by Crushing Sticks -- recipe will
change drastically once Mechanical Power is added. Currently only used
for efficient crafting of Tannics.
- Added all Tanning recipes. To go from Hide to Leather, players must
place the Hide on the Rack, right-click it with salt to cure it,
right-click the Cured Hide with any Dagger to Scud it, then right-click
the Scudded Hide with a bottle of Tannic and wait for the Leather to
tan.
- Each type of hide has a different base quality, but the quality of
Dagger used, as well as the Quality of Tannic will affect the quality of
the final piece of Leather.
- There are 4 possible Leather outcomes -- Botched Leather, Suede,
Leather, and Fine Leather in order of quality.
- Recipes requiring Leather now require one of these 4, but will accept
any Leather of equal or higher quality as well (ie, a Book only needs
Suede, but can be crafted with Fine Leather if desired).
- Changed recipes for all items crafted with Leather to match the new
system. Simple items like Books only require Suede, whereas quality
items like Wing Segments require Fine Leather. Botched Leather is
currently useless, as Suede is the lowest tier used for crafting
anything.
- Added 4 new Hide items -- Horse Hide, Pig Skin, Wolf Hide, and Polar
Bear Pelts. Dropped from their respective mobs.
- Vanilla Leather has been renamed and retextured to Cow Hide.
- Flesh and Rotten Flesh can both also be tanned into Leather, though of
much lower base quality than any of the other skins.
- Added a ton of in-between items for Leatherworking, though most of
them won't need to be obtained since the entire process can be completed
without removing the Hide from the Rack.
- Added Jerky! Comes in every type of meat, including Cod and Salmon.
- Jerky is tanned with its respective raw meat and a piece of Salt.
Takes 30 seconds to dry and requires sun. Jerky heals 1 food shenk less
than its cooked meat counterpart, but restores 2x as much saturation,
making it great for regenerating health without wasting hunger points.
- Added Tannic -- a vegetable solution required for tanning Hide into
Leather. Has an enormous amount of potential crafting recipes, and the
ingredients used will affect the quality of the final product. Sawdust
and Poison Potatoes are the most effective ingredients, but nearly any
other plant-based item can be used as well.
- Changed Potion of Infection to require an Awkward base.
- Added recipe for Potion of Harmony.
- 500000 fucking json files.