H A L L O W E M E
-Added Shimmerstone Geodes! Created when a Shimmerstone source souches
Lava or Water. No uses currently, will be fully implemented next update.
- Fixed a boatload of Skull bugs.
- Added all 20 Mob Skulls! Dropped at a rate of 1/20 during late
October, and 1/666 the rest of the year.
- Added Shimmering Ore! New ore found exclusively in the Nether.
Generates under Lava oceans, and emits particles from the top so you can
see it even when it's totally submerged. Can be alloyed into a Bucket of
Crystalfloe (will be switched over to the Melter when we re-add that).
- Minestrapp Skulls can now be worn on your head.
- Fixed potential world-tick event crash with CrystalFloe.
- Added Shimmering Chunks. Obtained by crushing Shimmering Ore. Can be
alloyed into Crystalfloe Buckets.