SPIKES!!! HOW DO YOU LIKE MY SPIKES!!!
- Added Spikes -- a game-changing, multipurpose automation block. So far
only the Wood ones are in-game, but more are coming soon.
- When pushed by a Piston, Spikes will break any block in front of them
(provided it could be harvested with a tool of the same material as the
Spike).
- Spikes deal damage to mobs, players, and vehicles that touch the
pointy end, making them great for automated mob grinders.
- Damage dealt by Spikes is based on not only the Spike's material, but
also how the entity collides with it. Simply touching the Spike will
deal damage equivallent to a Dagger of that material, while falling on
one will deal damage similar to a Sword, and pushing a Spike into a mob
with a Piston will deal damage similar to a Mace.
- Added Copper, Tin, Bronze, Steel, Torite, Archantine, Glacierite,
Blazium, and Dimensium Nuggets.
- Added Iron, Torite, Diamond, Archantine, Glacierite, Blazium, and
Dimensium Plating.
- Renamed and retextured Titanium to Archantine, a new fantasy metal, so
as to reduce game balance conflicts with mods that add a more realistic
interpretation of Titanium.
- Fixed Crates not saving their inventory contents when the world is
closed.