Irresponsible Luminescence!
- Added Carpet Glow Moss! New moss variant that will eventually generate
in Overworld caves. Has a model similar to a Carpet, but with side
textures that hang over the block below it. Behaves very similar to
normal Moss, but emits a low light level and glows in the dark.
- Added Creeping Glow Moss! New moss variant that will generate rarely
in every dimension. Behaves similar to Moss, but upside-down, requiring
a block above it to survive. Has hair-like tendrils of varying lengths
and side textures that appear if it is adjacent to a solid block. Emits
a low light level and glows in the dark.
- Added Glowing Dye. Comes in 9 varieties: White, Magenta, Red, Orange,
Gold, Green, Cyan, Blue, and Purple. Not currently usable or craftable.
- Fixed Moss only being able to survive at completely random coordinates
in the Nether.
- Fixed Berry Bushes dropping way too many Berries when broken while
fully grown.
- Fixed a crash that could occurr while placing an item with no bonus
drop in a Crusher while the Crusher's bonus slot is occupied.
- Added a bunch of textures for upcoming features.