Cheese Louise
- Added Cheesemaking Baskets! Crafted with a Basket and Tallow, or
alternatively created by just right-clicking a Basket with Tallow.
- Cheesemaking Baskets are (unsurprisingly) used for making Cheese.
Applying items to one in a specific order will complete the cheesemaking
process, which rewards the player with a full Wheel of Cheese.
- Wheels of Cheese can be right-clicked to eat a slice (like Cake) or
right-clicked with a Dagger to cut off a Slice of Cheese item.
- Added Cork Blocks! Used for cheesemaking, currently uncraftable.
- Added a shitload of food items. Most notably:
- Quesadillas! Alloyed with Cheese and Tortilla. Can be placed and eaten
directly or sliced with a Dagger.
- Pizza! Expensive food block that can be eaten directly or sliced with
a Dagger.
- Cake Slices! Cake can now be sliced with a Dagger as well, yielding a
Cake Slice item that can be taken and eaten on the go.
- Specific foods will now grant players the Well Fed effect when eaten!
This effect buffs movement speed, swing speed, and knockback resistance,
and stacks with other potion effects that give similar bonuses. The
duration and effectiveness is dependent on the food type, and it only
applies if the player is at full hunger after the food item is consumed.
- Added Coconuts! Grown from Palm Trees. They slowly mature at the top
of the tree before breaking off and falling to the ground. Drops Coconut
Slices when broken, or can be harvested with Silk Touch and buried in
the Sand to grow another Palm Tree.
- Added Coconut Slices! Food item that returns an empty Coconut Shell
when eaten. The slices can be crafted into Milk, and the Shells can be
smelted into Charcoal or crafted into Primitive Armor.
- Added Palm Fronds! Decorative leaf block generated by Palm Trees. Can
be placed facing 16 different directions. Drops Dead Palm Fronds unless
harvested with Shears.
- Added Palm Crowns! Decorative wood blocks that generate at the top of
Palm Trees. Palm Crowns will generate additional Coconuts over time, but
only if the ground is clear of them, there are Palm Fronds above each
viable Coconut location, and the Palm Crown was naturally generated, not
harvested by a player.
- Added Palm Logs, Planks, etcetera! New Wood type! Not that much else
to say about those!
- Added the Pressurizer! Major WIP, but it sort of functions.
- Probably some other stuff I forgot!